refactoring: adjust contract of luaT_tuple_new()
Accept an index on the Lua stack. Report an error using the diagnostics
area (not raise a Lua error).

The function is used by built-in key_def and merger modules on
tarantool-2.*. Now we plan to expose necessary APIs from tarantool of
all supported versions (including 1.10) and create external key_def and
merger modules.

Part of #5273

(cherry picked from commit 64a6464d31c92ea11c4f5c3ccdaada2ae9666455)

The original commit message:

 | lua: add luaT_tuple_new()
 |
 | The function allows to create a tuple with specific tuple format in C
 | code using a Lua table, another tuple, or objects on a Lua stack.
 |
 | Needed for #3276, #3398, #4025

The original message is a bit misleading, because it was not rewritten
after introduction of the function with the same name in
9e2a905c069e58cdd95d91868025391ecd9404e6 ('box: fix format of tuple
produced with frommap()').