tarantoolctl fails to stop server
When server process couldn't be stopped with SIGTERM as described in
[1], then tarantoolctl didn't check its status and returned success
exit code. To fix it, check of the process after SIGTERM send should
be added and if the process was still alive than it should be killed
with SIGKILL. Also some time delay added before the process status
check to be sure that it completely gone.

Closed #5245

[1] - https://github.com/tarantool/tarantool/issues/5244