test-run: restart server on each test in worker
Found that non restartable workers, used to run groups of tasks, had
only single tarantool server process started in the very start of it
at the worker's initalization. If any test in the worker's queue was
aborted then all the rest tests in queue couldn't be started. To fix
it the server restart commands were added just after the each task
finish. This change could decrease the performance, but it did not,
because in parallel runs these servers restarts don't use resources
and in absolute testing time was not increased. It happened because
there were other bottle necks in resources that made this change
invisible for the performance. Checking the change w/o parallel runs
found that even in this mode the running time wasn't increased too.

Closes tarantool/test-run#192
Closes tarantool/test-run#137
Closes tarantool/test-run#215