WIP: module api/lua: add luaL_iscdata() function
It is useful to provide a module specific error when cdata expected, but
a value of another type is passed.

Alternative would be using of lua_type() to check against LUA_TCDATA,
but this constant is not exposed for modules. See more in the
luaL_iscdata() API comment.

XXX: Add a test to test/app-tap/module_api.{c,test.lua}.

Part of #5273