WIP: module api/lua: add luaT_tuple_encode()
It is the same as luaT_tuple_new(), but returns raw MsgPack data (not
<box_tuple_t>) allocated on Lua shared ibuf (<tarantool_lua_ibuf>).

The reason to introduce this function is to provide ability to use
box_tuple_compare_with_key() function from an external module for a key
passed as a Lua table. The compare function has the following signature:

 | API_EXPORT int
 | box_tuple_compare_with_key(box_tuple_t *tuple_a, const char *key_b,
 |                            box_key_def_t *key_def);

The second parameter is a key encoded as an MsgPack array, not a tuple
structure. So luaT_tuple_new() is not applicable here (it is not
worthful to create a tuple structure if we need just MsgPack data).

XXX: Add a module API test.

Part of #5273

(cherry picked from commit 5cbb2e41bb32a3184d2663f678b705aef823dc85)